+++ src/controllers/ProductsController.php
@@ -4,25 +4,15 @@ namespace craft\digitalproducts\controllers;
 
 use Craft;
 use craft\base\Element;
-use craft\base\Field;
 use craft\digitalproducts\elements\License;
 use craft\digitalproducts\elements\Product;
-use craft\digitalproducts\models\ProductType;
 use craft\digitalproducts\Plugin as DigitalProducts;
 use craft\digitalproducts\web\assets\cp\Bundle;
-use craft\errors\ElementNotFoundException;
-use craft\errors\MissingComponentException;
-use craft\helpers\DateTimeHelper;
-use craft\helpers\Json;
-use craft\helpers\Localization;
+use craft\helpers\Cp;
 use craft\helpers\UrlHelper;
-use craft\models\Site;
 use craft\web\Controller as BaseController;
-use craft\web\UrlManager;
-use DateTime;
 use Throwable;
 use yii\base\Exception;
-use yii\base\InvalidConfigException;
 use yii\web\BadRequestHttpException;
 use yii\web\ForbiddenHttpException;
 use yii\web\Response;
@@ -52,90 +42,71 @@ class ProductsController extends BaseController
 
     /**
      * Index of digital products
-     *
-     * @return Response
-     * @throws InvalidConfigException
      */
     public function actionIndex(): Response
     {
-        $editableProductTypes = \craft\digitalproducts\Plugin::getInstance()->getProductTypes()->getEditableProductTypes();
+        $editableProductTypes = DigitalProducts::getInstance()->getProductTypes()->getEditableProductTypes();
         $this->getView()->registerAssetBundle(Bundle::class);
         return $this->renderTemplate('digital-products/products/index', compact('editableProductTypes'));
     }
 
     /**
-     * Edit a product
+     * Create a new Product and redirect to its edit page.
      *
-     * @param string $productTypeHandle the product type handle
-     * @param int|null $productId the product id
-     * @param string|null $siteHandle the site handle
-     * @param Product|null $product the product
-     * @return Response
-     * @throws Exception in case of a missing product type or an incorrect site handle.
+     * @throws BadRequestHttpException
+     * @throws ForbiddenHttpException
      */
-    public function actionEdit(string $productTypeHandle, int $productId = null, string $siteHandle = null, Product $product = null): Response
+    public function actionCreate(string $productTypeHandle): ?Response
     {
-        $productType = null;
-
-        $variables = [
-            'productTypeHandle' => $productTypeHandle,
-            'productId' => $productId,
-            'product' => $product,
-        ];
-
-        // Make sure a correct product type handle was passed so we can check permissions
-        if ($productTypeHandle) {
-            $productType = DigitalProducts::getInstance()->getProductTypes()->getProductTypeByHandle($productTypeHandle);
-        }
+        $productType = DigitalProducts::getInstance()->getProductTypes()->getProductTypeByHandle($productTypeHandle);
 
         if (!$productType) {
-            throw new Exception('The product type was not found.');
+            throw new BadRequestHttpException("Invalid product type handle: $productTypeHandle");
         }
 
         $this->requirePermission('digitalProducts-manageProductType:' . $productType->uid);
-        $variables['productType'] = $productType;
 
-        if ($siteHandle !== null) {
-            $variables['site'] = Craft::$app->getSites()->getSiteByHandle($siteHandle);
-
-            if (!$variables['site']) {
-                throw new Exception('Invalid site handle: ' . $siteHandle);
-            }
+        $site = Cp::requestedSite();
+        if (!$site) {
+            throw new ForbiddenHttpException('User not authorized to edit content in any sites.');
         }
 
-        $this->_prepareVariableArray($variables);
+        $product = Craft::createObject(Product::class);
+        $product->typeId = $productType->id;
+        $product->siteId = $site->id;
 
-        /** @var Product $product */
-        $product = $variables['product'];
-        if (!empty($product->id)) {
-            $variables['title'] = $product->title;
-        } else {
-            $variables['title'] = Craft::t('digital-products', 'Create a new product');
+        $user = static::currentUser();
+        if (!Craft::$app->getElements()->canSave($product, $user)) {
+            throw new ForbiddenHttpException('User not authorized to create a product.');
         }
 
-        // Can't just use the entry's getCpEditUrl() because that might include the site handle when we don't want it
-        $variables['baseCpEditUrl'] = 'digital-products/products/' . $variables['productTypeHandle'] . '/{id}';
+        $product->setScenario(Element::SCENARIO_ESSENTIALS);
+        $success = Craft::$app->getDrafts()->saveElementAsDraft($product, $user->id, markAsSaved: false);
 
-        // Set the "Continue Editing" URL
-        $variables['continueEditingUrl'] = $variables['baseCpEditUrl'] .
-            (Craft::$app->getIsMultiSite() && Craft::$app->getSites()->currentSite->id !== $variables['site']->id ? '/' . $variables['site']->handle : '');
+        if (!$success) {
+            return $this->asModelFailure($product, Craft::t('digital-products', 'Couldn\'t create product.'), 'product');
+        }
 
-        $this->_maybeEnableLivePreview($variables);
+        $editUrl = $product->getCpEditUrl();
 
-        //$variables['promotions']['sales'] = Commerce::getInstance()->getSales()->getSalesForProduct($variables['product']);
+        $response = $this->asModelSuccess($product, Craft::t('digital-products', 'Product created.'), 'product', [
+            'cpEditUrl' => $this->request->getIsCpRequest() ? $editUrl : null,
+        ]);
 
-        return $this->renderTemplate('digital-products/products/_edit', $variables);
+        if (!$this->request->getAcceptsJson()) {
+            $response->redirect(UrlHelper::urlWithParams($editUrl, ['fresh' => 1]));
+        }
+
+        return $response;
     }
 
     /**
      * Delete a product.
      *
-     * @return Response|null
      * @throws BadRequestHttpException
      * @throws Exception if no product found
      * @throws ForbiddenHttpException
      * @throws Throwable
-     * @throws MissingComponentException
      */
     public function actionDeleteProduct(): ?Response
     {
@@ -149,10 +120,9 @@ class ProductsController extends BaseController
             ->one();
 
         if (!$product) {
-            throw new Exception(Craft::t('digital-products', 'No product exists with the ID “{id}”.', ['id' => $productId]));
+            throw new Exception(Craft::t('digital-products', 'No product exists with the ID "{id}".', ['id' => $productId]));
         }
 
-        /** @var ProductType $productType */
         $productType = $product->getType();
         $this->requirePermission('digitalProducts-manageProductType:' . $productType->uid);
 
@@ -160,23 +130,20 @@ class ProductsController extends BaseController
 
         if ($licensesExist || !Craft::$app->getElements()->deleteElement($product)) {
             if (Craft::$app->getRequest()->getAcceptsJson()) {
-                $this->asJson(['success' => false]);
+                return $this->asJson(['success' => false]);
             }
 
-            $error = $licensesExist ? Craft::t('digital-products', 'Couldn’t delete product with existing licenses.') : Craft::t('digital-products', 'Couldn’t delete product.');
+            $error = $licensesExist
+                ? Craft::t('digital-products', 'Couldn\'t delete product with existing licenses.')
+                : Craft::t('digital-products', 'Couldn\'t delete product.');
 
             Craft::$app->getSession()->setError($error);
-            /** @var UrlManager $urlManager */
-            $urlManager = Craft::$app->getUrlManager();
-            $urlManager->setRouteParams([
-                'product' => $product,
-            ]);
 
             return null;
         }
 
         if (Craft::$app->getRequest()->getAcceptsJson()) {
-            $this->asJson(['success' => true]);
+            return $this->asJson(['success' => true]);
         }
 
         Craft::$app->getSession()->setNotice(Craft::t('digital-products', 'Product deleted.'));
@@ -185,53 +152,8 @@ class ProductsController extends BaseController
     }
 
     /**
-     * Save a new or an existing product.
-     *
-     * @return Response|null
-     * @throws BadRequestHttpException
-     * @throws Exception
-     * @throws ForbiddenHttpException
-     * @throws Throwable
-     * @throws ElementNotFoundException
-     * @throws MissingComponentException
-     */
-    public function actionSave(): ?Response
-    {
-        $this->requirePostRequest();
-
-        $product = $this->_buildProductFromPost();
-
-        $this->requirePermission('digitalProducts-manageProductType:' . $product->getType()->uid);
-
-        $existingProduct = (bool)$product->id;
-
-        // Always use live scenario until we implement temporary SKUs
-        $product->setScenario(Element::SCENARIO_LIVE);
-
-        if (!Craft::$app->getElements()->saveElement($product)) {
-            if (!$existingProduct) {
-                $product->id = null;
-            }
-
-            Craft::$app->getSession()->setError(Craft::t('app', 'Couldn’t save product.'));
-            /** @var UrlManager $urlManager */
-            $urlManager = Craft::$app->getUrlManager();
-            $urlManager->setRouteParams([
-                'product' => $product,
-            ]);
-
-            return null;
-        }
-
-        Craft::$app->getSession()->setNotice(Craft::t('app', 'Product saved.'));
-
-        return $this->redirectToPostedUrl($product);
-    }
-
-    /**
      * Previews a product.
      *
-     * @return Response
      * @throws BadRequestHttpException
      * @throws Exception
      * @throws ForbiddenHttpException
@@ -241,7 +163,16 @@ class ProductsController extends BaseController
     {
         $this->requirePostRequest();
 
-        $product = $this->_buildProductFromPost();
+        $productId = Craft::$app->getRequest()->getBodyParam('productId');
+        $siteId = Craft::$app->getRequest()->getBodyParam('siteId');
+
+        /** @var Product|null $product */
+        $product = Craft::$app->getElements()->getElementById($productId, Product::class, $siteId);
+
+        if (!$product) {
+            throw new BadRequestHttpException('Invalid product ID: ' . $productId);
+        }
+
         $this->requirePermission('digitalProducts-manageProductType:' . $product->getType()->uid);
 
         return $this->_showProduct($product);
@@ -250,11 +181,8 @@ class ProductsController extends BaseController
     /**
      * Redirects the client to a URL for viewing a disabled product on the front end.
      *
-     * @param int $productId the product id
-     * @param int|null $siteId the site id
-     *
-     * @return Response
-     * @throws Exception if there's no valid product template
+     * @throws BadRequestHttpException
+     * @throws ServerErrorHttpException
      */
     public function actionShareProduct(int $productId, int $siteId = null): Response
     {
@@ -271,7 +199,6 @@ class ProductsController extends BaseController
 
         $this->requirePermission('digitalProducts-manageProductType:' . $product->getType()->uid);
 
-        // Create the token and redirect to the product URL with the token in place
         $token = Craft::$app->getTokens()->createToken([
             'action' => 'digital-products/products/viewSharedProduct',
             'params' => ['productId' => $productId, 'site' => $product->getSite()],
@@ -285,10 +212,6 @@ class ProductsController extends BaseController
     /**
      * Shows a product/draft/version based on a token.
      *
-     * @param int $productId
-     * @param int|null $siteId
-     *
-     * @return Response
      * @throws Exception if product not found
      */
     public function actionViewSharedProduct(int $productId, ?int $siteId = null): Response
@@ -308,9 +231,6 @@ class ProductsController extends BaseController
     /**
      * Displays a product.
      *
-     * @param Product $product
-     *
-     * @return Response
      * @throws Exception if product type is not found
      */
     private function _showProduct(Product $product): Response
@@ -335,7 +255,6 @@ class ProductsController extends BaseController
 
         Craft::$app->language = $site->language;
 
-        // Have this product override any freshly queried products with the same ID/site
         if ($product->id) {
             Craft::$app->getElements()->setPlaceholderElement($product);
         }
@@ -345,193 +264,5 @@ class ProductsController extends BaseController
         return $this->renderTemplate($siteSettings[$product->siteId]->template, [
             'product' => $product,
         ]);
-    }
-
-    /**
-     * Prepare $variable array for editing a Product
-     *
-     * @param array $variables by reference
-     *
-     * @throws ForbiddenHttpException if missing permissions
-     * @throws Exception if data ir missing or corrupt
-     */
-    private function _prepareVariableArray(array &$variables): void
-    {
-        // Locale related checks
-        if (Craft::$app->getIsMultiSite()) {
-            // Only use the sites that the user has access to
-            $variables['siteIds'] = Craft::$app->getSites()->getEditableSiteIds();
-        } else {
-            $variables['siteIds'] = [Craft::$app->getSites()->getPrimarySite()->id];
-        }
-
-        if (!$variables['siteIds']) {
-            throw new ForbiddenHttpException('User not permitted to edit content in any sites supported by this section');
-        }
-
-        if (empty($variables['site'])) {
-            $site = $variables['site'] = Craft::$app->getSites()->currentSite;
-
-            if (!in_array($variables['site']->id, $variables['siteIds'], false)) {
-                $site = $variables['site'] = Craft::$app->getSites()->getSiteById($variables['siteIds'][0]);
-            }
-        } else {
-            // Make sure they were requesting a valid site
-            /** @var Site $site */
-            $site = $variables['site'];
-            if (!in_array($site->id, $variables['siteIds'], false)) {
-                throw new ForbiddenHttpException('User not permitted to edit content in this site');
-            }
-        }
-
-        // Product related checks
-        /** @var ProductType $productType */
-        $productType = $variables['productType'];
-
-        if (empty($variables['product'])) {
-            if (!empty($variables['productId'])) {
-                $variables['product'] = Craft::$app->getElements()->getElementById($variables['productId'], Product::class, $site->id);
-
-                if (!$variables['product']) {
-                    throw new Exception('Missing product data.');
-                }
-            } else {
-                $variables['product'] = new Product();
-                $variables['product']->typeId = $productType->id;
-
-                if (!empty($variables['siteId'])) {
-                    $variables['product']->site = $variables['siteId'];
-                }
-            }
-        }
-
-        /** @var Product $product */
-        $product = $variables['product'];
-        $form = $productType->getFieldLayout()->createForm($product);
-        $variables['tabs'] = $form->getTabMenu();
-        $variables['fieldsHtml'] = $form->render();
-
-        // Enable locales
-        if ($variables['product']->id) {
-            $variables['enabledSiteIds'] = Craft::$app->getElements()->getEnabledSiteIdsForElement($variables['product']->id);
-        } else {
-            $variables['enabledSiteIds'] = [];
-
-            foreach (Craft::$app->getSites()->getEditableSiteIds() as $site) {
-                $variables['enabledSiteIds'][] = $site;
-            }
-        }
-
-        foreach ($productType->getProductFieldLayout()->getTabs() as $index => $tab) {
-            // Do any of the fields on this tab have errors?
-            $hasErrors = false;
-
-            if ($product->hasErrors()) {
-                foreach ($tab->getLayout()->getCustomFields() as $field) {
-                    /** @var Field $field */
-                    if ($hasErrors = $product->hasErrors($field->handle . '.*')) {
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Enable live preview for products with valid templates on desktop browsers.
-     *
-     * @param array $variables
-     * @throws Exception
-     * @throws InvalidConfigException
-     */
-    private function _maybeEnableLivePreview(array &$variables): void
-    {
-        if (
-            !Craft::$app->getRequest()->isMobileBrowser(true) &&
-            DigitalProducts::getInstance()->getProductTypes()->isProductTypeTemplateValid($variables['productType'], $variables['product']->siteId)
-        ) {
-            $this->getView()->registerJs('Craft.LivePreview.init(' . Json::encode([
-                    'fields' => '#title-field, #fields > div > div > .field, #sku-field, #price-field',
-                    'extraFields' => '#meta-pane .field',
-                    'previewUrl' => $variables['product']->getUrl(),
-                    'previewAction' => Craft::$app->getSecurity()->hashData('digital-products/products/preview-product'),
-                    'previewParams' => [
-                        'typeId' => $variables['productType']->id,
-                        'productId' => $variables['product']->id,
-                        'siteId' => $variables['product']->siteId,
-                    ],
-                ]) . ');');
-
-            $variables['showPreviewBtn'] = true;
-
-            // Should we show the Share button too?
-            if ($variables['product']->id) {
-                // If the product is enabled, use its main URL as its share URL.
-                if ($variables['product']->getStatus() === Product::STATUS_LIVE) {
-                    $variables['shareUrl'] = $variables['product']->getUrl();
-                } else {
-                    $variables['shareUrl'] = UrlHelper::actionUrl('digital-products/products/share-product', [
-                        'productId' => $variables['product']->id,
-                        'siteId' => $variables['product']->siteId,
-                    ]);
-                }
-            }
-        } else {
-            $variables['showPreviewBtn'] = false;
-        }
-    }
-
-    /**
-     * Build product from POST data.
-     *
-     * @return Product
-     * @throws Exception if product cannot be found
-     * @throws Throwable
-     */
-    private function _buildProductFromPost(): Product
-    {
-        $request = Craft::$app->getRequest();
-        $productId = $request->getParam('productId');
-        $site = $request->getParam('siteId');
-
-        if ($productId) {
-            /** @var Product|null $product */
-            $product = Craft::$app->getElements()->getElementById($productId, Product::class, $site);
-
-            if (!$product) {
-                throw new Exception('No product found with that id.');
-            }
-        } else {
-            $product = new Product();
-        }
-
-        $product->typeId = $request->getBodyParam('typeId');
-        $product->enabled = (bool) $request->getBodyParam('enabled');
-
-        $product->basePrice = (float)Localization::normalizeNumber($request->getBodyParam('basePrice', 0));
-        $product->setSku($request->getBodyParam('sku'));
-
-        $product->postDate = (($date = $request->getParam('postDate')) !== false ? (DateTimeHelper::toDateTime($date) ?: null) : $product->postDate);
-        $product->expiryDate = (($date = $request->getParam('expiryDate')) !== false ? (DateTimeHelper::toDateTime($date) ?: null) : $product->expiryDate);
-
-        $product->promotable = (bool)$request->getBodyParam('promotable');
-        $product->taxCategoryId = $request->getBodyParam('taxCategoryId');
-        $product->slug = $request->getBodyParam('slug');
-
-        $product->enabledForSite = (bool)$request->getBodyParam('enabledForSite', $product->enabledForSite);
-        $product->title = $request->getBodyParam('title', $product->title);
-        $product->setFieldValuesFromRequest('fields');
-
-        // Last checks
-        if (empty($product->getSku())) {
-            $productType = $product->getType();
-            $product->setSku(Craft::$app->getView()->renderObjectTemplate($productType->skuFormat, $product));
-        }
-
-        if (!$product->postDate) {
-            $product->postDate = new DateTime();
-        }
-
-        return $product;
     }
 }