+++ src/plugin/Routes.php
@@ -22,13 +22,13 @@ trait Routes
 
                 $event->rules['digital-products/products'] = 'digital-products/products/index';
                 $event->rules['digital-products/products/<productTypeHandle:{handle}>'] = 'digital-products/products/index';
-                $event->rules['digital-products/products/<productTypeHandle:{handle}>/new'] = 'digital-products/products/edit';
-                $event->rules['digital-products/products/<productTypeHandle:{handle}>/new/<siteHandle:{handle}>'] = 'digital-products/products/edit';
-                $event->rules['digital-products/products/<productTypeHandle:{handle}>/<productId:\d+>'] = 'digital-products/products/edit';
-                $event->rules['digital-products/products/<productTypeHandle:{handle}>/<productId:\d+>/<siteHandle:{handle}>'] = 'digital-products/products/edit';
+                $event->rules['digital-products/products/<productTypeHandle:{handle}>/new'] = 'digital-products/products/create';
+                $event->rules['digital-products/products/<productTypeHandle:{handle}>/<elementId:\d+><slug:(?:-[^\/]*)?>'] = 'elements/edit';
 
-                $event->rules['digital-products/licenses/new'] = 'digital-products/licenses/edit';
-                $event->rules['digital-products/licenses/<licenseId:\d+>'] = 'digital-products/licenses/edit';
+                $event->rules['digital-products/licenses/new'] = 'digital-products/licenses/create';
+                $event->rules['digital-products/licenses/<elementId:\d+>'] = 'elements/edit';
+
+                $event->rules['digital-products/settings/licenses'] = 'digital-products/settings/edit-licenses';
             }
         );
     }